requirements.txt
alembic==1.3.0
altgraph==0.16.1
app==0.0.1
appdirs==1.4.3
asn1crypto==1.2.0
astroid==2.3.2
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.213
botocore==1.12.213
bottle==0.12.17
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cheroot==6.5.6
CherryPy==18.1.2
Click==7.0
colorama==0.4.1
constantly==15.1.0
coverage==4.5.4
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.6
django-appconf==1.0.3
django-autoslug==1.9.6
django-compressor==2.3
django-debug-toolbar==2.0
django-haystack==2.8.1
django-ipware==2.1.0
django-mdeditor==0.1.16
django-uuslug==1.1.9
docopt==0.6.2
docutils==0.15.2
dominate==2.4.0
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
entrypoints==0.3
et-xmlfile==1.0.1
feedparser==5.2.1
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.17.1
gensim==3.8.0
guess-language-spirit==0.5.3
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jaraco.functools==2.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.39
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonpickle==1.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
langid==1.1.6
lazy-object-proxy==1.4.2
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
markdown2==2.3.8
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mysqlclient==1.4.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
notebook==6.0.1
numpy==1.17.0
olefile==0.46
openpyxl==3.0.0
ordered-set==3.1.1
packaging==19.2
pandas==0.25.0
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
Pattern==3.6
pdfkit==0.6.1
pdfminer.six==20181108
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.2.1
pipreqs==0.4.9
portend==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyBrain==0.3
pycparser==2.19
pycryptodome==3.8.2
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyInstaller==3.5
PyLaTeX==1.3.1
pylint==2.4.3
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.6.5
PyQt5==5.12
PyQt5-sip==4.19.19
pyrsistent==0.15.4
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.3
python-editor==1.0.4
python-igraph==0.7.1.post6
python-logstash==0.4.6
python-memcached==1.59
python-slugify==4.0.0
pytz==2019.3
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.3
queuelib==1.5.0
raven==6.10.0
RBTools==1.0.2
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.7.3
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
smart-open==1.8.4
sortedcontainers==2.0.5
soupsieve==1.9.3
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.34.2
sqlparse==0.3.0
tempora==1.14.1
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
textblob==0.15.3
texttable==1.6.2
Theano==1.0.4
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
Twisted==19.7.0
typed-ast==1.4.0
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.6
visitor==0.1.3
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
WeRoBot==1.10.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
xmltodict==0.12.0
yarg==0.1.9
zc.lockfile==2.0
zope.interface==4.6.0
